# docs/specs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WhoopCredentialsBody": {
        "description": "The client_id / client_secret pasted from developer.whoop.com.\n\nWHOOP client ids/secrets are short opaque strings; the 512-char cap is\ngenerous headroom without letting a hostile caller send a megabyte.",
        "properties": {
          "client_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Client Id",
            "type": "string"
          },
          "client_secret": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Client Secret",
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "client_secret"
        ],
        "title": "WhoopCredentialsBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Private local memory — notes, mails, messages, and docs",
    "title": "Estormi MCP",
    "version": "0.0.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/whoop/auth/callback": {
      "get": {
        "description": "Browser-redirect handler for the WHOOP OAuth flow.",
        "operationId": "whoop_auth_callback_api_whoop_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Whoop Auth Callback"
      }
    },
...
    "/api/whoop/credentials/upload": {
      "post": {
        "operationId": "whoop_credentials_upload_api_whoop_credentials_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhoopCredentialsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Whoop Credentials Upload"
      }
    },
...
  }
}